convertToPng/requirements.txt
Pillow>=10.0.0
pytest>=7.0.0  # Optional, if you prefer pytest for running tests
